Add Companion Guide and refactor companion modules
- Added a new Companion Guide to the documentation for better user guidance.
- Refactored several companion modules to improve type hinting and code clarity, including the addition of return types for methods.
- Enhanced the CompanionBase class with new methods for managing contacts and advertisements, improving overall functionality.
- Updated constants and models to support new features and ensure consistency across the codebase.
