+++ src/pymc_core/companion/models.py
@@ -1,5 +1,7 @@
 """Data models for companion radio state objects."""
 
+from __future__ import annotations
+
 from dataclasses import dataclass, field
 from typing import Optional
