+++ src/pymc_core/protocol/constants.py
@@ -66,6 +66,7 @@ TIMESTAMP_SIZE = 4  # 4 bytes for a timestamp (32-bit unsigned int)
 ADVERT_FLAG_IS_CHAT_NODE = 0x01
 ADVERT_FLAG_IS_REPEATER = 0x02
 ADVERT_FLAG_IS_ROOM_SERVER = 0x03
+ADVERT_FLAG_IS_SENSOR = 0x04
 ADVERT_FLAG_HAS_LOCATION = 0x10
 ADVERT_FLAG_HAS_FEATURE1 = 0x20
 ADVERT_FLAG_HAS_FEATURE2 = 0x40
